Make the get_columns_in_relation function case-insensitive for the pg adapter
From the dbt Slack: https://getdbt.slack.com/archives/C030A0UF5LM/p1648011287283189

```I'm confused about the intent of the get_columns_in_relation macro with respect to the case-sensitivity of the input relation argument: the postgres impl uses the information_schema.columns query construct, which is case-sensitive-- but the snowflake impl uses the describe table construct, which is not case-sensitive. AFAICT, it looks like this inconsistent behavior across adapters has been around for awhile and so it hasn't been causing anyone too much trouble, but it caused me some confusion when I was porting over some Snowflake-oriented dbt SQL to run against DuckDB and the dbt_utils.star macro failed in one of my models b/c it was silently relying on the case-insensitive behavior of the Snowflake adapter.```

It seems like the cleanest/easiest fix to enforce consistent behavior across the core adapters is to change up this macro to use Postgres' `LOWER` function when matching table names/schemas.